// src/app/modules/ticket/interfaces/ticket-print.ts
export interface TicketPrint {
    EventName: string;
    EventCode: string;
    VenueName: string;
    EventDate: string;
    EventFormattedDate: string;
    RefCode: string;
    CategoryName: string;
    SeatInfo: string;
    SeatAllocationType: number;
    Barcode: string;
    TicketBottomDescription: string;
    GateCode: string;
    GateName: string;
    BlockName: string;
    RowNumber: string;
    SeatName: string;
}